+++ packages/snap-preact/components/src/components/Molecules/Swatches/Swatches.tsx
@@ -1,10 +1,9 @@
 import { h } from 'preact';
 import classnames from 'classnames';
 import { css } from '@emotion/react';
-import { Theme, useTheme, CacheProvider, useTreePath, useSnap } from '../../../providers';
+import { Theme, useTheme, CacheProvider, useTreePath } from '../../../providers';
 import { ComponentProps, SwatchOption, BreakpointsProps, StyleScript } from '../../../types';
-import { useA11y, useComponent, useDisplaySettings } from '../../../hooks';
-import type { SnapTemplates } from '../../../../../src';
+import { useA11y, useDisplaySettings, useCustomComponentOverride } from '../../../hooks';
 import { defined, mergeProps, mergeStyles } from '../../../utilities';
 import { Grid, GridProps } from '../Grid';
 import { ImageProps, Image } from '../../Atoms/Image';
@@ -75,7 +74,6 @@ const defaultStyles: StyleScript<SwatchesProps> = ({ theme }) => {
 
 export function Swatches(properties: SwatchesProps) {
 	const globalTheme: Theme = useTheme();
-	const snap = useSnap();
 	const globalTreePath = useTreePath();
 
 	const defaultCarouselBreakpoints = {
@@ -123,14 +121,12 @@ export function Swatches(properties: SwatchesProps) {
 		};
 	}
 
-	const { onSelect, disabled, options, hideLabels, disableStyles, className, internalClassName, type, slideshow, grid, treePath, customComponent } =
-		props;
+	const { onSelect, disabled, options, hideLabels, disableStyles, className, internalClassName, type, slideshow, grid, treePath } = props;
 
-	if (customComponent) {
-		const ComponentOverride = useComponent((snap as SnapTemplates)?.templates?.library.import.component.swatches || {}, customComponent);
-		if (ComponentOverride) {
-			return <ComponentOverride {...props} />;
-		}
+	const { overrideElement, shouldRenderDefault } = useCustomComponentOverride('swatches', props);
+
+	if (!shouldRenderDefault) {
+		return overrideElement;
 	}
 
 	const subProps: SwatchesSubProps = {