+++ packages/snap-preact/components/src/themes/pike/components/templates/autocompleteModal.ts
@@ -82,7 +82,7 @@ export const autocompleteModal: ThemeComponent<'autocompleteModal', Autocomplete
 		autocompleteModal: {
 			...(autocompleteModalThemeComponentProps.default?.['autocompleteModal'] || {}),
 			themeStyleScript: autocompleteModalStyleScript,
-			layout: 'standard',
+			layout: 'desktop',
 			width: '90vw',
 			contentTitle: 'Product Suggestions',
 		},
@@ -137,7 +137,7 @@ export const autocompleteModal: ThemeComponent<'autocompleteModal', Autocomplete
 		...autocompleteModalThemeComponentProps.mobile,
 		autocompleteModal: {
 			...(autocompleteModalThemeComponentProps.mobile?.['autocompleteModal'] || {}),
-			layout: 'mini',
+			layout: 'mobile',
 		},
 		'autocompleteModal terms': {
 			vertical: false,
@@ -159,7 +159,7 @@ export const autocompleteModal: ThemeComponent<'autocompleteModal', Autocomplete
 		...autocompleteModalThemeComponentProps.tablet,
 		autocompleteModal: {
 			...(autocompleteModalThemeComponentProps.tablet?.['autocompleteModal'] || {}),
-			layout: 'standard',
+			layout: 'desktop',
 		},
 		'autocompleteModal terms': {
 			vertical: false,
@@ -197,7 +197,7 @@ export const autocompleteModal: ThemeComponent<'autocompleteModal', Autocomplete
 		...autocompleteModalThemeComponentProps.desktop,
 		autocompleteModal: {
 			...(autocompleteModalThemeComponentProps.desktop?.['autocompleteModal'] || {}),
-			layout: 'standard',
+			layout: 'desktop',
 		},
 		'autocompleteModal terms': {
 			vertical: false,